+++ app/src/main/java/com/example/pintube/ui/mypage/MypageViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.pintube.ui.notifications
+package com.example.pintube.ui.mypage
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class MypageViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
