+++ app/src/main/java/com/example/pintube/ui/shorts/ShortsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.pintube.ui.dashboard
+package com.example.pintube.ui.shorts
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.pintube.databinding.FragmentDashboardBinding
+import com.example.pintube.databinding.FragmentShortsBinding
 
-class DashboardFragment : Fragment() {
+class ShortsFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentShortsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val shortsViewModel =
+            ViewModelProvider(this).get(ShortsViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentShortsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
+        shortsViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
