+++ app/src/main/java/com/example/pintube/ui/shorts/ShortsViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.pintube.ui.dashboard
+package com.example.pintube.ui.shorts
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ShortsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
